pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reverse_distillation"
version = "0.1.0"
description = "Reverse Distillation for Protein Language Models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"accelerate==1.7.0",
	"bio==1.8.0",
	"biopython==1.85",
	"biotite==1.2.0",
	"biotraj==1.2.2",
	"Brotli==1.1.0",
	"datasets",
	"decorator==5.2.1",
	"dotenv==0.9.9",
	"einops==0.8.1",
	"fair-esm==2.0.0",
	"fbpca==1.0",
	"filelock==3.18.0",
	"fsspec==2025.3.2",
	"gprofiler-official==1.0.0",
	"h5py==3.13.0",
	"huggingface-hub==0.32.3",
	"hydra-core==1.3.2",
	"hydra-optuna-sweeper==1.2.0",
	"idna==3.10",
	"ipykernel>=7.0.1",
	"ipython==8.36.0",
	"loguru>=0.7.3",
	"memory-profiler==0.61.0",
	"mygene==3.2.2",
	"networkx==3.4.2",
	"numpy==2.2.5",
	"nvidia-cublas-cu12==12.6.4.1; sys_platform == 'linux'",
	"nvidia-cuda-cupti-cu12==12.6.80; sys_platform == 'linux'",
	"nvidia-cuda-nvrtc-cu12==12.6.77; sys_platform == 'linux'",
	"nvidia-cuda-runtime-cu12==12.6.77; sys_platform == 'linux'",
	"nvidia-cudnn-cu12==9.5.1.17; sys_platform == 'linux'",
	"nvidia-cufft-cu12==11.3.0.4; sys_platform == 'linux'",
	"nvidia-cufile-cu12==1.11.1.6; sys_platform == 'linux'",
	"nvidia-curand-cu12==10.3.7.77; sys_platform == 'linux'",
	"nvidia-cusolver-cu12==11.7.1.2; sys_platform == 'linux'",
	"nvidia-cusparse-cu12==12.5.4.2; sys_platform == 'linux'",
	"nvidia-cusparselt-cu12==0.6.3; sys_platform == 'linux'",
	"nvidia-nccl-cu12==2.26.2; sys_platform == 'linux'",
	"nvidia-nvjitlink-cu12==12.6.85; sys_platform == 'linux'",
	"nvidia-nvtx-cu12==12.6.77; sys_platform == 'linux'",
	"omegaconf==2.3.0",
	"optuna==2.10.1",
	"pandas==2.2.2",
	"peft==0.15.2",
	"python-dateutil",
	"python-dotenv==1.1.0",
	"PyYAML==6.0.2",
	"ruff==0.14.0",
	"safetensors==0.5.3",
	"scikit-learn==1.6.1",
	"scipy==1.15.3",
	"tokenizers==0.21.1",
	"torch==2.7.0",
	"torchtext==0.18.0",
	"torchvision==0.22.0",
	"tqdm==4.67.1",
	"transformers==4.48.1",
	"typing_extensions==4.13.2",
	"urllib3==2.4.0",
]


[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
	"pre-commit",
]


[tool.ruff]
line-length = 90
target-version = "py312"
include = [
	"src/reverse_distillation/**",
]


[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src", "reverse_distillation"]

# Specify that packages live in the src directory
[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
]
